swap from click to mousemove on demo
I originally thought that moving the mouse would make this thing work and was surprised when it didn't. So I made it work in a slightly more expected manner (imo).

While doing so I noticed that the yellow circle was not centered on the mouse position so that got fixed as well.